committor test
