+++ tests/markov/tools/analysis/test_committor.py
@@ -4,64 +4,35 @@ r"""Unit tests for the committor API-function
 
 """
 
-import unittest
 import numpy as np
+import pytest
 
 from deeptime.data import birth_death_chain
+from deeptime.markov.tools.analysis import committor
 from tests.markov.tools.numeric import assert_allclose
 
-from deeptime.markov.tools.analysis import committor
+
+@pytest.fixture
+def bdc(sparse_mode):
+    p = np.zeros(10)
+    q = np.zeros(10)
+    p[0:-1] = 0.5
+    q[1:] = 0.5
+    p[4] = 0.01
+    q[6] = 0.1
+    bdc = birth_death_chain(q, p, sparse=sparse_mode)
+    yield bdc
 
 
-class TestCommittorDense(unittest.TestCase):
-    def setUp(self):
-        p = np.zeros(10)
-        q = np.zeros(10)
-        p[0:-1] = 0.5
-        q[1:] = 0.5
-        p[4] = 0.01
-        q[6] = 0.1
-
-        self.bdc = birth_death_chain(q, p)
-
-    def tearDown(self):
-        pass
-
-    def test_forward_comittor(self):
-        P = self.bdc.transition_matrix
-        un = committor(P, [0, 1], [8, 9], forward=True)
-        u = self.bdc.committor_forward(1, 8)
-        assert_allclose(un, u)
-
-    def test_backward_comittor(self):
-        P = self.bdc.transition_matrix
-        un = committor(P, [0, 1], [8, 9], forward=False)
-        u = self.bdc.committor_backward(1, 8)
-        assert_allclose(un, u)
+def test_forward_committor(bdc):
+    P = bdc.transition_matrix
+    un = committor(P, [0, 1], [8, 9], forward=True)
+    u = bdc.committor_forward(1, 8)
+    assert_allclose(un, u)
 
 
-class TestCommittorSparse(unittest.TestCase):
-    def setUp(self):
-        p = np.zeros(100)
-        q = np.zeros(100)
-        p[0:-1] = 0.5
-        q[1:] = 0.5
-        p[49] = 0.01
-        q[51] = 0.1
-
-        self.bdc = birth_death_chain(q, p)
-
-    def tearDown(self):
-        pass
-
-    def test_forward_comittor(self):
-        P = self.bdc.transition_matrix_sparse
-        un = committor(P, list(range(10)), list(range(90, 100)), forward=True)
-        u = self.bdc.committor_forward(9, 90)
-        assert_allclose(un, u)
-
-    def test_backward_comittor(self):
-        P = self.bdc.transition_matrix_sparse
-        un = committor(P, list(range(10)), list(range(90, 100)), forward=False)
-        u = self.bdc.committor_backward(9, 90)
-        assert_allclose(un, u)
+def test_backward_comittor(bdc):
+    P = bdc.transition_matrix
+    un = committor(P, [0, 1], [8, 9], forward=False)
+    u = bdc.committor_backward(1, 8)
+    assert_allclose(un, u)
